+++ test/lib/gem_bootstrap/test_gem_bootstrap.rb
@@ -1,9 +1,9 @@
-require_relative '../../test_helper'
-require 'fileutils'
+require_relative "../../test_helper"
+require "fileutils"
 describe GemBootstrap do
-  context '#dummy_test' do
-    it 'must pass the simple test' do
-      'string'.wont_be_nil
+  context "#dummy_test" do
+    it "must pass the simple test" do
+      "string".wont_be_nil
     end
   end
 end