+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-### Changelogs
+#### 0.1.2
+
+- Fix the gemspec
+- Remove the rubocop-todo.yml and use only the .rubocop.yml in the template
+- Misc code cleanup
+
+#### 0.1.1
+
+- Simplify the CLI interface
 
 #### 0.1.0